feat: various discord bot tweaks (#47)
* refactor: deploy command

* fix: disable report command

* fix: check deletable before deleting

* fix: don't assume guild exists

* feat: support channel partial

* feat: role based auth

* refactor: rename rooms command

* feat: reset global commands

* feat: convert view to use ephemeral

* feat: send booking confirmation

* chore: update deps

+++ src/discord/commands/manage.ts
@@ -1,7 +1,8 @@
-import { ButtonInteraction, CommandInteraction, Message, MessageActionRow, MessageButton, MessageEmbed, SelectMenuInteraction } from 'discord.js';
+import { ButtonInteraction, CommandInteraction, Message, MessageActionRow, MessageButton, MessageEmbed, SelectMenuInteraction, TextChannel } from 'discord.js';
 import { Types } from 'mongoose';
 import TimeblockModel from '../../models/timeBlock.model';
 import SectionModel from '../../models/section.model';
+import { getBookingInfoEmbed } from './view';
 
 function updateEmbed(userArray: string[], maxCapacity: number, manageState: string) {
     const embedTitle = manageState === 'add' ? 'Adding' : 'Removing';
@@ -180,18 +181,28 @@ export default {
         const message = (await interaction.reply({ content: `${interaction.user}`, embeds: [embed], components: [buttonRow], fetchReply: true })) as Message;
 
         const buttonCollector = message.createMessageComponentCollector({ componentType: 'BUTTON', time: 120000 });
-        const mentionsCollector = interaction!.channel!.createMessageCollector({ time: 120000 });
+        const channel = interaction.channel?.partial ? await interaction.channel.fetch() : (interaction.channel as TextChannel);
+        const mentionsCollector = channel!.createMessageCollector({ time: 120000 });
 
         buttonCollector.on('collect', async (buttonInteraction: ButtonInteraction) => {
             //Temporary check as message isn't ephemeral
             if (buttonInteraction.user.id === interaction.user.id) {
                 switch (buttonInteraction.customId) {
-                    case 'completeBooking':
+                    case 'completeBooking': {
                         await TimeblockModel.updateOne({ _id: bookingId }, { $push: { users: await mongoDBFilter(userArray, bookingId) } }, { upsert: true });
 
-                        interaction.followUp({ content: 'Booking Successfully Booked!', ephemeral: true });
-                        interaction.deleteReply();
+                        const infoEmbed = await getBookingInfoEmbed(buttonInteraction.client, bookingId);
+                        infoEmbed.setTitle('Booking Confirmation');
+                        try {
+                            await interaction.user.send({ embeds: [infoEmbed] });
+                            interaction.followUp({ content: "Booking Successfully Booked! We've sent you a confirmation in your DMs.", ephemeral: true });
+                            interaction.deleteReply();
+                        } catch (e) {
+                            interaction.followUp({ embeds: [infoEmbed], ephemeral: true });
+                            interaction.deleteReply();
+                        }
                         break;
+                    }
                     case 'addCollaborators':
                     case 'removeCollaborators':
                         manageState = buttonInteraction.customId === 'addCollaborators' ? 'add' : 'remove';
@@ -216,7 +227,7 @@ export default {
                     return;
                 }
 
-                message.delete();
+                if (message.deletable) message.delete();
 
                 for (const mentionedUser of mentionedUsers) {
                     switch (manageState) {
